small two patches for win32, still some problems on win32 (with kde)
svn path=/trunk/KDE/kdeedu/marble/; revision=732143

+++ src/lib/PlaceMarkContainer.cpp
@@ -27,7 +27,7 @@ inline bool populationLessThan( GeoDataPlaceMark* mark1, GeoDataPlaceMark* mark2
 { 
     // If compared items do not differ in terms of being selected,
     // compare them based on population numbers.
-    return ( abs( mark1->popularity() ) > abs( mark2->popularity() ) );
+    return ( abs( (double) mark1->popularity() ) > abs( (double) mark2->popularity() ) );
 }
